Fix deploy: migrate middleware.ts to proxy.ts for Next.js 16
Next.js 16 replaces middleware.ts with proxy.ts (export named 'proxy').
Having both files causes a build error. Rewrote proxy.ts with correct
auth logic using our auth_token cookie + verifyToken from lib/auth,
replacing the incorrect NextAuth-based version. Deleted middleware.ts.
